refactor: export add announcement button to separate component
